+++ Sources/SwiftDOF/Parser/ASCII.swift
@@ -11,8 +11,8 @@ enum ASCII {
   @usableFromInline static let slash: UInt8 = 0x2F  // '/'
   @usableFromInline static let zero: UInt8 = 0x30  // '0'
   @usableFromInline static let nine: UInt8 = 0x39  // '9'
-  @usableFromInline static let A: UInt8 = 0x41
   @usableFromInline static let E: UInt8 = 0x45
+  @usableFromInline static let I: UInt8 = 0x49
   @usableFromInline static let N: UInt8 = 0x4E
   @usableFromInline static let O: UInt8 = 0x4F
   @usableFromInline static let S: UInt8 = 0x53